feat: support an instance of BeautifulSoup as input
